requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
async-timeout==5.0.1
bcrypt==5.0.0
billiard==4.2.4
blinker==1.9.0
brotli==1.2.0
celery==5.6.2
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
ConfigArgParse==1.7.1
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.0
Flask==3.1.2
flask-cors==6.0.2
Flask-Login==0.6.3
flower==2.0.1
gevent==25.9.1
geventhttpclient==2.3.7
greenlet==3.2.4
h11==0.16.0
humanize==4.13.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.6.2
locust==2.34.0
MarkupSafe==3.0.3
msgpack==1.1.2
opentelemetry-api==1.39.1
opentelemetry-exporter-prometheus==0.60b1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
packaging==25.0
passlib==1.7.4
prometheus_client==0.24.0
prompt_toolkit==3.0.52
psutil==7.2.1
pyasn1==0.6.1
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose==3.5.0
pytz==2025.2
pyzmq==27.1.0
redis==7.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
starlette==0.49.3
tomli==2.4.0
tornado==6.5.4
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.39.0
vine==5.1.0
wcwidth==0.2.14
Werkzeug==3.1.5
wrapt==1.17.3
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1
